requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
langchain
openai
python-dotenv
pydantic
unstructured==0.7.12
websockets
metaphor_python